--- hellion-core/src/main/java/org/monarchinitiative/hellion/core/noise/PhenopacketNoise.java
@@ -1,12 +1,0 @@
-package org.monarchinitiative.hellion.core.noise;
-
-import org.phenopackets.schema.v2.Phenopacket;
-
-/**
- * Implementors add noise to the provided phenopacket.
- */
-public interface PhenopacketNoise {
-
-    Phenopacket distort(Phenopacket pp);
-
-}
