+++ phenoimp-core/src/main/java/org/monarchinitiative/phenoimp/core/DistortionRunner.java
@@ -1,0 +1,18 @@
+package org.monarchinitiative.phenoimp.core;
+
+import org.phenopackets.schema.v2.Phenopacket;
+
+import java.nio.file.Path;
+
+/**
+ * Implementors can distort the provided phenopacket.
+ */
+public interface DistortionRunner {
+
+    static DistortionRunnerBuilder builder(Path dataDirectory) {
+        return new DistortionRunnerBuilder(dataDirectory);
+    }
+
+    Phenopacket run(Phenopacket pp);
+
+}
