+++ phenoimp-core/src/main/java/org/monarchinitiative/phenoimp/core/noise/PhenopacketNoise.java
@@ -1,0 +1,12 @@
+package org.monarchinitiative.phenoimp.core.noise;
+
+import org.phenopackets.schema.v2.Phenopacket;
+
+/**
+ * Implementors add noise to the provided phenopacket.
+ */
+public interface PhenopacketNoise {
+
+    Phenopacket distort(Phenopacket pp);
+
+}
